-small package refactoring according to Artem's comments
