+++ src/main/java/com/breskul/bibernate/annotation/GeneratedValue.java
@@ -1,5 +1,7 @@
 package com.breskul.bibernate.annotation;
 
+import com.breskul.bibernate.annotation.enums.Strategy;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
