+++ src/main/java/com/breskul/bibernate/annotation/enums/Strategy.java
@@ -1,4 +1,4 @@
-package com.breskul.bibernate.annotation;
+package com.breskul.bibernate.annotation.enums;
 
 public enum Strategy {
 	SEQUENCE, IDENTITY, AUTO
