+++ src/test/java/com/breskul/bibernate/persistence/EntityManagerImplTest.java
@@ -6,9 +6,9 @@ import com.breskul.bibernate.exception.JdbcDaoException;
 import com.breskul.bibernate.exception.LazyInitializationException;
 import com.breskul.bibernate.exception.TransactionException;
 import com.breskul.bibernate.persistence.test_model.*;
-import com.breskul.bibernate.persistence.testmodel.cascadepersist.CompanyCascadePersist;
-import com.breskul.bibernate.persistence.testmodel.cascadepersist.NoteComplexCascadePersist;
-import com.breskul.bibernate.persistence.testmodel.cascadepersist.PersonCascadePersist;
+import com.breskul.bibernate.persistence.test_model.cascadepersist.CompanyCascadePersist;
+import com.breskul.bibernate.persistence.test_model.cascadepersist.NoteComplexCascadePersist;
+import com.breskul.bibernate.persistence.test_model.cascadepersist.PersonCascadePersist;
 import com.breskul.bibernate.persistence.util.DaoUtils;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityTransaction;
@@ -21,7 +21,6 @@ import java.sql.SQLException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.Month;
-import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
