+++ src/test/java/com/breskul/bibernate/persistence/test_model/Company.java
@@ -4,7 +4,7 @@ import com.breskul.bibernate.annotation.*;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import static com.breskul.bibernate.annotation.Strategy.SEQUENCE;
+import static com.breskul.bibernate.annotation.enums.Strategy.SEQUENCE;
 
 @Entity
 @Data
