+++ src/test/java/com/breskul/bibernate/persistence/test_model/PersonWithoutEntity.java
@@ -3,7 +3,7 @@ package com.breskul.bibernate.persistence.test_model;
 import com.breskul.bibernate.annotation.Column;
 import com.breskul.bibernate.annotation.GeneratedValue;
 import com.breskul.bibernate.annotation.Id;
-import com.breskul.bibernate.annotation.Strategy;
+import com.breskul.bibernate.annotation.enums.Strategy;
 import lombok.Data;
 
 import java.time.LocalDate;
