+++ src/test/java/com/breskul/bibernate/persistence/test_model/cascadepersist/PersonCascadePersist.java
@@ -1,6 +1,8 @@
-package com.breskul.bibernate.persistence.testmodel.cascadepersist;
+package com.breskul.bibernate.persistence.test_model.cascadepersist;
 
 import com.breskul.bibernate.annotation.*;
+import com.breskul.bibernate.annotation.enums.CascadeType;
+import com.breskul.bibernate.annotation.enums.Strategy;
 import lombok.Data;
 import lombok.ToString;
