+++ src/test/java/com/breskul/bibernate/persistence/util/test_model/EntityIdentity.java
@@ -2,7 +2,7 @@ package com.breskul.bibernate.persistence.util.test_model;
 
 import com.breskul.bibernate.annotation.GeneratedValue;
 import com.breskul.bibernate.annotation.Id;
-import com.breskul.bibernate.annotation.Strategy;
+import com.breskul.bibernate.annotation.enums.Strategy;
 
 public class EntityIdentity {
     @Id
